# src/modules/users/dto/verify-email.dto.ts
import { IsNotEmpty, IsString, Length, IsEmail } from 'class-validator';
import { ApiProperty } from '@nestjs/swagger';

export class VerifyEmailDto {
  @ApiProperty({
    example: 'test@example.com',
    description: 'User email address',
  })
  @IsString()
  @IsNotEmpty()
  @IsEmail()
  email: string;

  @ApiProperty({
    example: 'A1B2C3',
    description: 'Verification code',
  })
  @IsString()
  @IsNotEmpty()
  @Length(5)
  code: string;
}